feat: Add conditional structures for user login
This commit adds conditional structures to the JavaScript code for user login. It prompts the user for their ID and age, and if the age is greater than or equal to 18 and the ID length is 11, it displays a success message. Otherwise, it displays an error message and redirects the user to Google's homepage.